Put MemIRFile in its own file
